refactor: use faceit production api endpoints
since the extension is running on the client side of a logged in user,
there is no need to use a developer api. Developer API is limited to 10k
requests per hour, which introduces an issue for large amounts of users.
Use of production endpoints with each user's own existing bearer token
lets the requests come from each user separately rather then use a
single key generated on the faceit developer portal.

This commit also includes the updated typings for the affected api endpoints.
In addition to adjusting the typings to the structure of production api responses, I
decided not to type data points that are not used by the extension's
codebase. Datapoints needed for potential new features can be typed
during the development of said features.